Make things compile again

+++ src/deh_mapping.cpp
@@ -140,7 +140,7 @@ bool DEH_SetStringMapping(deh_context_t *context, deh_mapping_t *mapping,
 
     // Copy value into field:
 
-    M_StringCopy(location, value, entry->size);
+    M_StringCopy(location, value, static_cast<size_t>(entry->size));
 
     return true;
 }
