+++ src/doom/p_pspr.cpp
@@ -512,7 +512,7 @@ void A_Punch(mobj_t *, player_t *player, pspdef_t *)
         damage *= 10;
 
     angle = player->mo->angle;
-    angle += P_SubRandom() << 18;
+    angle += static_cast<unsigned int>(P_SubRandom() << 18);
     slope = P_AimLineAttack(player->mo, angle, MELEERANGE);
     P_LineAttack(player->mo, angle, MELEERANGE, slope, damage);
 
@@ -540,7 +540,7 @@ void A_Saw(mobj_t *, player_t *player, pspdef_t *)
     if (!player) return; // [crispy] let pspr action pointers get called from mobj states
     damage = 2 * (P_Random() % 10 + 1);
     angle  = player->mo->angle;
-    angle += P_SubRandom() << 18;
+    angle += static_cast<unsigned int>(P_SubRandom() << 18);
 
     // use meleerange + 1 se the puff doesn't skip the flash
     slope = P_AimLineAttack(player->mo, angle, MELEERANGE + 1);
@@ -691,7 +691,7 @@ void P_GunShot(mobj_t *mo,
     angle  = mo->angle;
 
     if (!accurate)
-        angle += P_SubRandom() << 18;
+        angle += static_cast<unsigned int>(P_SubRandom() << 18);
 
     P_LineAttack(mo, angle, MISSILERANGE, bulletslope, damage);
 }
@@ -771,7 +771,7 @@ void A_FireShotgun2(mobj_t *, player_t *player, pspdef_t *)
     {
         damage = 5 * (P_Random() % 3 + 1);
         angle  = player->mo->angle;
-        angle += P_SubRandom() << ANGLETOFINESHIFT;
+        angle += static_cast<unsigned int>(P_SubRandom() << ANGLETOFINESHIFT);
         P_LineAttack(player->mo,
             angle,
             MISSILERANGE,
