+++ src/doom/p_setup.cpp
@@ -143,7 +143,7 @@ void P_LoadVertexes(int lump)
     numvertexes = static_cast<int>(W_LumpLength(lump) / sizeof(mapvertex_t));
 
     // Allocate zone memory for buffer.
-    vertexes = zmalloc<decltype(vertexes)>(numvertexes * sizeof(vertex_t), PU_LEVEL, 0);
+    vertexes = zmalloc<decltype(vertexes)>(static_cast<unsigned long>(numvertexes) * sizeof(vertex_t), PU_LEVEL, 0);
 
     // Load data into cache.
     data = cache_lump_num<uint8_t *>(lump, PU_STATIC);
@@ -202,8 +202,8 @@ void P_LoadSegs(int lump)
     int       sidenum;
 
     numsegs = static_cast<int>(W_LumpLength(lump) / sizeof(mapseg_t));
-    segs    = zmalloc<decltype(segs)>(numsegs * sizeof(seg_t), PU_LEVEL, 0);
-    memset(segs, 0, numsegs * sizeof(seg_t));
+    segs    = zmalloc<decltype(segs)>(static_cast<unsigned long>(numsegs) * sizeof(seg_t), PU_LEVEL, 0);
+    memset(segs, 0, static_cast<unsigned long>(numsegs) * sizeof(seg_t));
     data = cache_lump_num<uint8_t *>(lump, PU_STATIC);
 
     ml = reinterpret_cast<mapseg_t *>(data);
@@ -317,7 +317,7 @@ void P_LoadSubsectors(int lump)
     subsector_t *   ss;
 
     numsubsectors = static_cast<int>(W_LumpLength(lump) / sizeof(mapsubsector_t));
-    subsectors    = zmalloc<decltype(subsectors)>(numsubsectors * sizeof(subsector_t), PU_LEVEL, 0);
+    subsectors    = zmalloc<decltype(subsectors)>(static_cast<unsigned long>(numsubsectors) * sizeof(subsector_t), PU_LEVEL, 0);
     data          = cache_lump_num<uint8_t *>(lump, PU_STATIC);
 
     // [crispy] fail on missing subsectors
@@ -325,7 +325,7 @@ void P_LoadSubsectors(int lump)
         I_Error("P_LoadSubsectors: No subsectors in map!");
 
     ms = reinterpret_cast<mapsubsector_t *>(data);
-    memset(subsectors, 0, numsubsectors * sizeof(subsector_t));
+    memset(subsectors, 0, static_cast<unsigned long>(numsubsectors) * sizeof(subsector_t));
     ss = subsectors;
 
     for (i = 0; i < numsubsectors; i++, ss++, ms++)
@@ -353,8 +353,8 @@ void P_LoadSectors(int lump)
         I_Error("P_LoadSectors: No sectors in map!");
 
     numsectors = static_cast<int>(W_LumpLength(lump) / sizeof(mapsector_t));
-    sectors    = zmalloc<decltype(sectors)>(numsectors * sizeof(sector_t), PU_LEVEL, 0);
-    memset(sectors, 0, numsectors * sizeof(sector_t));
+    sectors    = zmalloc<decltype(sectors)>(static_cast<unsigned long>(numsectors) * sizeof(sector_t), PU_LEVEL, 0);
+    memset(sectors, 0, static_cast<unsigned long>(numsectors) * sizeof(sector_t));
     data = cache_lump_num<uint8_t *>(lump, PU_STATIC);
 
     // [crispy] fail on missing sectors
@@ -403,7 +403,7 @@ void P_LoadNodes(int lump)
     node_t *   no;
 
     numnodes = static_cast<int>(W_LumpLength(lump) / sizeof(mapnode_t));
-    nodes    = zmalloc<decltype(nodes)>(numnodes * sizeof(node_t), PU_LEVEL, 0);
+    nodes    = zmalloc<decltype(nodes)>(static_cast<unsigned long>(numnodes) * sizeof(node_t), PU_LEVEL, 0);
     data     = cache_lump_num<uint8_t *>(lump, PU_STATIC);
 
     // [crispy] warn about missing nodes
@@ -534,8 +534,8 @@ void P_LoadLineDefs(int lump)
     int           warn, warn2; // [crispy] warn about invalid linedefs
 
     numlines = static_cast<int>(W_LumpLength(lump) / sizeof(maplinedef_t));
-    lines    = zmalloc<decltype(lines)>(numlines * sizeof(line_t), PU_LEVEL, 0);
-    memset(lines, 0, numlines * sizeof(line_t));
+    lines    = zmalloc<decltype(lines)>(static_cast<unsigned long>(numlines) * sizeof(line_t), PU_LEVEL, 0);
+    memset(lines, 0, static_cast<unsigned long>(numlines) * sizeof(line_t));
     data = cache_lump_num<uint8_t *>(lump, PU_STATIC);
 
     mld  = reinterpret_cast<maplinedef_t *>(data);
@@ -676,8 +676,8 @@ void P_LoadSideDefs(int lump)
     side_t *      sd;
 
     numsides = static_cast<int>(W_LumpLength(lump) / sizeof(mapsidedef_t));
-    sides    = zmalloc<decltype(sides)>(numsides * sizeof(side_t), PU_LEVEL, 0);
-    memset(sides, 0, numsides * sizeof(side_t));
+    sides    = zmalloc<decltype(sides)>(static_cast<unsigned long>(numsides) * sizeof(side_t), PU_LEVEL, 0);
+    memset(sides, 0, static_cast<unsigned long>(numsides) * sizeof(side_t));
     data = cache_lump_num<uint8_t *>(lump, PU_STATIC);
 
     msd = reinterpret_cast<mapsidedef_t *>(data);
@@ -718,7 +718,7 @@ bool P_LoadBlockMap(int lump)
     // adapted from boom202s/P_SETUP.C:1025-1076
     wadblockmaplump = zmalloc<decltype(wadblockmaplump)>(lumplen, PU_LEVEL, nullptr);
     W_ReadLump(lump, wadblockmaplump);
-    blockmaplump = zmalloc<decltype(blockmaplump)>(sizeof(*blockmaplump) * count, PU_LEVEL, nullptr);
+    blockmaplump = zmalloc<decltype(blockmaplump)>(sizeof(*blockmaplump) * static_cast<unsigned long>(count), PU_LEVEL, nullptr);
     blockmap     = blockmaplump + 4;
 
     blockmaplump[0] = SHORT(wadblockmaplump[0]);
@@ -747,7 +747,7 @@ bool P_LoadBlockMap(int lump)
 
     count      = static_cast<int>(sizeof(*blocklinks)) * bmapwidth * bmapheight;
     blocklinks = zmalloc<decltype(blocklinks)>(count, PU_LEVEL, 0);
-    memset(blocklinks, 0, count);
+    memset(blocklinks, 0, static_cast<size_t>(count));
 
     // [crispy] (re-)create BLOCKMAP if necessary
     fprintf(stderr, ")\n");
@@ -796,7 +796,7 @@ void P_GroupLines()
     }
 
     // build line tables for each sector
-    linebuffer = zmalloc<decltype(linebuffer)>(totallines * sizeof(line_t *), PU_LEVEL, 0);
+    linebuffer = zmalloc<decltype(linebuffer)>(static_cast<unsigned long>(totallines) * sizeof(line_t *), PU_LEVEL, 0);
 
     for (i = 0; i < numsectors; ++i)
     {
@@ -944,7 +944,7 @@ static void PadRejectArray(uint8_t *array, unsigned int len)
         0x1d4a11 // DOOM_CONST_ZONEID
     };
 
-    rejectpad[0] = ((totallines * 4 + 3) & ~3) + 24;
+    rejectpad[0] = static_cast<unsigned int>(((totallines * 4 + 3) & ~3) + 24);
 
     // Copy values from rejectpad into the destination array.
 
@@ -976,7 +976,7 @@ static void PadRejectArray(uint8_t *array, unsigned int len)
             padvalue = 0xf00;
         }
 
-        memset(array + sizeof(rejectpad), padvalue, len - sizeof(rejectpad));
+        memset(array + sizeof(rejectpad), static_cast<int>(padvalue), len - sizeof(rejectpad));
     }
 }
 
@@ -984,7 +984,7 @@ static void P_LoadReject(int lumpnum)
 {
     // Calculate the size that the REJECT lump *should* be.
 
-    size_t minlength = (numsectors * numsectors + 7) / 8;
+    size_t minlength = static_cast<size_t>((numsectors * numsectors + 7) / 8);
 
     // If the lump meets the minimum length, it can be loaded directly.
     // Otherwise, we need to allocate a buffer of the correct size
@@ -1001,7 +1001,7 @@ static void P_LoadReject(int lumpnum)
         rejectmatrix = zmalloc<decltype(rejectmatrix)>(minlength, PU_LEVEL, &rejectmatrix);
         W_ReadLump(lumpnum, rejectmatrix);
 
-        PadRejectArray(rejectmatrix + lumplen, static_cast<int>(minlength - lumplen));
+        PadRejectArray(rejectmatrix + lumplen, static_cast<unsigned int>(minlength - lumplen));
     }
 }
