+++ src/doom/p_sight.cpp
@@ -293,12 +293,12 @@ bool P_CrossBSPNode(int bspnum)
     node_t *bsp;
     int     side;
 
-    if (bspnum & NF_SUBSECTOR)
+    if (static_cast<unsigned int>(bspnum) & NF_SUBSECTOR)
     {
         if (bspnum == -1)
             return P_CrossSubsector(0);
         else
-            return P_CrossSubsector(bspnum & (~NF_SUBSECTOR));
+            return P_CrossSubsector(static_cast<unsigned int>(bspnum) & (~NF_SUBSECTOR));
     }
 
     bsp = &nodes[bspnum];
